Test: 서비스 분리로 인한 예산등록 테스트코드 수정
